requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tensorflow>=2.6.0
pyyaml>=5.4.0
joblib>=1.0.0
pyvisa>=1.11.0  # For hardware interface
pyserial>=3.5  # For serial communication
tkinter  # Usually comes with Python 